End of the day update
Have completed the code until the ICA component saving stage. The dual regression concept is completed. Need to include a saving function that saves each of the components in the patient specific directory.  Also, uploading a script that serves as my script for trying out all the functions apriori before finalization.